[format] support parquet reader reading field with 'FIXED_LEN_BYTE_ARRAY' type
